chore(pre-commit): apply hook fixes
